Rename auth route folder
