+++ src/api/v1/auth/routes/logout.ts
