+++ src/api/v1/index.ts
@@ -3,13 +3,13 @@ import { FastifyInstance, FastifyPluginOptions } from 'fastify'
 import employees from './employees'
 
 import permissions from './permissions'
-import authorization from './authorization'
+import auth from './auth'
 import register from './register'
 
 export default async (fastify: FastifyInstance, options: FastifyPluginOptions): Promise<void> => {
   await fastify
     .register(permissions, { prefix: '/permissions' })
-    .register(authorization, { prefix: '/auth' })
+    .register(auth, { prefix: '/auth' })
     .register(register, { prefix: '/register' })
     .register(employees, { prefix: '/employees' })
 }
